[cherry-pick 2.0-beta][Dy2Stat] Transforme api 'to_tensor' to 'assign'.  (#26873) (#27055)
